DEV/clinic/requirements.txt
annotated-types==0.6.0
antlr4-python3-runtime==4.13.1
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
aws-sam-translator==1.89.0
aws-xray-sdk==2.13.0
awsebcli==3.20.10
blessed==1.20.0
boto3==1.28.85
botocore==1.31.85
cement==2.8.2
certifi==2024.2.2
cffi==1.16.0
cfn-lint==0.87.3
channels==4.1.0
charset-normalizer==3.3.2
colorama==0.4.3
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
docker==7.0.0
graphql-core==3.2.3
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
joserfc==0.10.0
jschema-to-python==1.2.3
jsondiff==2.0.0
jsonpatch==1.33
jsonpath-ng==1.6.1
jsonpickle==3.0.4
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
junit-xml==1.9
lazy-object-proxy==1.10.0
Markdown==3.6
MarkupSafe==2.1.5
moto==5.0.7
mpmath==1.3.0
multipart==0.2.4
networkx==3.3
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==24.0
pathable==0.4.3
pathspec==0.10.1
pbr==6.0.0
pip-autoremove==0.10.0
pluggy==1.5.0
ply==3.11
psycopg2-binary==2.9.9
py-partiql-parser==0.5.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.31.1
regex==2024.5.15
requests==2.31.0
responses==0.25.0
rfc3339-validator==0.1.4
rpds-py==0.18.1
s3transfer==0.7.0
sarif-om==1.0.4
semantic-version==2.8.5
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
sympy==1.12
termcolor==1.1.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==1.26.5
wcwidth==0.1.9
Werkzeug==3.0.3
wrapt==1.16.0
xmltodict==0.13.0
zope.interface==6.4
